# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS School Management API
  version: 0.4.0
servers:
  - url: http://localhost:3001
paths:
  /students:
    get:
      summary: List students
      responses:
        '200':
          description: Students for the current tenant
    post:
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student created
  /students/{id}:
    get:
      summary: Get student
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Student
        '404':
          description: Not found
    put:
      summary: Update student
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student updated
    delete:
      summary: Delete student
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    StudentInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        classId:
          type: string
        admissionNumber:
          type: string
        parentContacts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              relationship:
                type: string
              phone:
                type: string
      required:
        - firstName
        - lastName
...